+++ deltacat/__init__.py
@@ -36,7 +36,6 @@ from deltacat.catalog.model.catalog import (  # noqa: F401
     get_catalog,
     put_catalog,
 )
-import deltacat.catalog.iceberg as IcebergCatalog
 from deltacat.catalog.model.table_definition import TableDefinition
 from deltacat.storage import (
     DistributedDataset,
