+++ deltacat/catalog/__init__.py
@@ -3,5 +3,12 @@ from deltacat.catalog.model.properties import (  # noqa: F401
     get_catalog_properties,
 )
 from deltacat.catalog.model.catalog import Catalog, Catalogs  # noqa: F401
+from deltacat.catalog.main import impl as DeltacatCatalog
 
-__all__ = ["CatalogProperties", "get_catalog_properties", "Catalog", "Catalogs"]
+__all__ = [
+    "CatalogProperties",
+    "get_catalog_properties",
+    "Catalog",
+    "Catalogs",
+    "DeltacatCatalog",
+]
