+++ requirements.txt
@@ -21,7 +21,7 @@ pydantic!=2.0.*,!=2.1.*,!=2.2.*,!=2.3.*,!=2.4.*,<3
 
 # deltacat[iceberg]
 #pyiceberg[glue] @ git+https://github.com/apache/iceberg-python
-pyiceberg[glue] == 0.8.0
+pyiceberg[glue] == 0.9.0
 pymemcache == 4.0.0
 ray[default] >= 2.20.0,<2.31.0
 redis == 4.6.0
